% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdf}
\alias{cdf}
\title{A function representing the CDF of the p1 estimates from Phillips et al. (2017).}
\format{An object of class \code{function} of length 1.}
\usage{
cdf
}
\description{
A function representing the CDF of the p1 estimates from Phillips et al. (2017).
}
\references{
Phillips et al., (2017). High sensitivity and interindividual variability
in the response of the human circadian system to evening light. PNAS.
}
\keyword{datasets}
