% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose-response.R
\name{cdf_inv_constructor}
\alias{cdf_inv_constructor}
\title{Constructed weighted inverse-cdf from a mixture of a cdf and a normal-cdf}
\usage{
cdf_inv_constructor(weight, cdf_full, mu, sigma, p1_vals)
}
\arguments{
\item{weight}{a value (0<=weight<=1) specifying weight given to cdf_full}

\item{cdf_full}{the cdf forming first component of mixture}

\item{mu}{the mean of the normal distribution component}

\item{sigma}{the sd of the normal distribution component}

\item{p1_vals}{a set of p1 values used to form mixture cdf}
}
\value{
a function yielding the inverse cdf for the mixture distribution
}
\description{
Takes the cdf of a distribution and combines it with a cdf of a N(mu, sigma)
in a weighted sum: cdf_mixed(x) = cdf(x) weight + normal_cdf(x|mu,sigma) (1-weight).
}
