% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{comparison_test}
\alias{comparison_test}
\title{Performs between- or within-individual experiments comparing melatonin suppression at two
lux levels using t tests}
\usage{
comparison_test(is_between, lux_1, lux_2, n, population_df, p_value = 0.05)
}
\arguments{
\item{is_between}{a Boolean indicating whether experiment is within or between type}

\item{lux_1}{first measured lux value}

\item{lux_2}{second measured lux value}

\item{n}{number of individuals in sample}

\item{population_df}{a tibble representing a virtual experiment}

\item{p_value}{a statistical test size}
}
\value{
a binary value indicating test success (if=1) or failure (if=0) to detect difference of correct sign
}
\description{
For a between-individual experiment, a t-test is used to compare the melatonin suppression level at two lux values for two different
samples of individuals. This function requires based an (ideally large) simulated population of individual dose-response data.
}
\details{
For a within-individual, a paired t-test is used to compare the melatonin suppression level at two lux values for a sample of individuals
measured at each of those values. This function requires based an (ideally large) simulated population of individual dose-response data.
}
\examples{
library(chronodoseresponse)

# generate virtual data for 200 individuals
population_df <- virtual_experiment(200)

# carry out 10 replicate between-individual experiments comparing lux_1=10 with lux_2=30
# for a test sample size of 20
is_between <- TRUE
results <- purrr::map_dbl(1:10, ~comparison_test(is_between, 10, 30, 20, population_df))

# calculate percentage of experiments obtaining the correct signed difference
mean(results)
# carry out 10 replicate within-individual experiments comparing lux_1=10 with lux_2=30
# for a test sample size of 20
is_between <- FALSE
results <- purrr::map_dbl(1:10, ~comparison_test(is_between, 10, 30, 20, population_df))

# calculate percentage of experiments obtaining the correct signed difference
mean(results)
}
