% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{comparison_test_single}
\alias{comparison_test_single}
\title{Performs a single between- or within-individual experiments comparing melatonin suppression at two
lux levels using t tests}
\usage{
comparison_test_single(is_between, lux_1, lux_2, n, population_df)
}
\arguments{
\item{is_between}{a Boolean indicating whether experiment is within or between type}

\item{lux_1}{first measured lux value}

\item{lux_2}{second measured lux value}

\item{n}{number of individuals in sample}

\item{population_df}{a tibble representing a virtual experiment}
}
\value{
a named list comprising 'result': a binary value indicating test success (if=1) or failure (if=0) to detect difference of correct sign;
and 'p_value': the p-value from the t test
}
\description{
For a between-individual experiment, a t-test is used to compare the melatonin suppression level at two lux values for two different
samples of individuals. This function requires based an (ideally large) simulated population of individual dose-response data.
}
\details{
For a within-individual, a paired t-test is used to compare the melatonin suppression level at two lux values for a sample of individuals
measured at each of those values. This function requires based an (ideally large) simulated population of individual dose-response data.
}
