% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{comparison_test_treatment}
\alias{comparison_test_treatment}
\title{Performs between- or within-individual experiments comparing melatonin suppression for
treated and untreated subgroups at a single lux}
\usage{
comparison_test_treatment(
  is_between,
  lux,
  n,
  population_treated_df,
  is_treated_higher,
  p_value = 0.05,
  nreps = 1
)
}
\arguments{
\item{is_between}{a Boolean indicating whether experiment is within or between type}

\item{lux}{the lux value at which to conduct the comparison}

\item{n}{number of individuals in sample}

\item{population_treated_df}{a virtual population data frame with individual who have been treated
and those who have been untreated (where treatment is indicated by a Boolean column named "treated")}

\item{is_treated_higher}{Boolean indicating whether treated group has a higher response}

\item{p_value}{a statistical test size}

\item{nreps}{integer indicating number of test replicates (which defaults to 1)}
}
\value{
a dataframe where each row corresponds to a test result in a replicate
}
\description{
For a between-individual comparison, a t-test comparing the melatonin suppression level
at a single lux is conducted on two subgroups (untreated and treated) comprising different
sets of individuals.
}
\details{
For a within-individual comparison, a paired t-test comparing the melatonin suppression level
at a single lux level for a set of individuals before and after they are treated is conducted.

This function requires based an (ideally large) simulated population of individual dose-response data.
}
\examples{
library(chronodoseresponse)
library(purrr)
library(dplyr)

# create a population where half are treated; half not
n <- 200
sample_untreated <- virtual_experiment(n) \%>\%
  mutate(treated=FALSE)
sample_treated <- virtual_experiment(n, treated_ed50_multiplier=0.5) \%>\%
  mutate(treated=TRUE)
population_df <- sample_untreated \%>\%
  bind_rows(sample_treated)

# carry out a within-study test at 10 lux for 30 people
# note that, because the ed50 multiplier for the treated group was
# <1, this group should have a higher response
is_treated_higher <- TRUE
comparison_test_treatment(FALSE, 10, 30, population_df, is_treated_higher)
}
