% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ed}
\alias{ed}
\title{Calculates ed for a given quantile}
\usage{
ed(edx, p1, p2)
}
\arguments{
\item{edx}{a quantile in \link{0, 1}}

\item{p1}{the log of the ed50 lux}

\item{p2}{a positive shape parameter}
}
\value{

}
\description{
Calculates ed for a given quantile
}
\examples{
# at ed50 the returned value is 10^p1
ed(0.5, 1.3, 4.4) # = 10^1.3
}
