% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose-response.R
\name{experimental_data}
\alias{experimental_data}
\title{Generates noisy data from a givn virtual population}
\usage{
experimental_data(
  sigma,
  simulated_pop,
  lux = c(10, 30, 50, 100, 200, 400, 2000)
)
}
\arguments{
\item{sigma}{a noise level}

\item{simulated_pop}{a tibble of logistic_2 parameters}

\item{lux}{a vector of lux values which defaults to c(10, 30, 50, 100, 200, 400, 2000)}
}
\value{
a tibble containing simulated noisy data
}
\description{
Generates noisy data from a givn virtual population
}
