% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f_cdf}
\alias{f_cdf}
\title{Calculates cumulative density for a given pdf}
\usage{
f_cdf(x, probability_density, lower)
}
\arguments{
\item{x}{a value}

\item{probability_density}{an integratable probability density function}

\item{lower}{a lower bound to integrate from}
}
\value{
a cdf value between 0 and 1
}
\description{
Calculates cumulative density for a given pdf
}
