% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f_sample_n}
\alias{f_sample_n}
\title{Generates n draws from underlying density using inverse transform sampling}
\usage{
f_sample_n(n, cdf_inv)
}
\arguments{
\item{n}{the number of draws}

\item{cdf_inv}{the inverse cumulative density function}
}
\value{
a vector of length n of draws
}
\description{
Generates n draws from underlying density using inverse transform sampling
}
