% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f_sample_one}
\alias{f_sample_one}
\title{Uses inverse transform sampling to generate an independent draw}
\usage{
f_sample_one(cdf_inv)
}
\arguments{
\item{cdf_inv}{an inverse cumulative density function}
}
\value{
a draw from underlying probability density
}
\description{
Uses inverse transform sampling to generate an independent draw
}
