% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{generate_two_samples}
\alias{generate_two_samples}
\title{Generates samples of measured melatonin values at two lux levels}
\usage{
generate_two_samples(is_between, lux_1, lux_2, n, population_df)
}
\arguments{
\item{is_between}{a Boolean indicating whether experiment is within or between type}

\item{lux_1}{first measured lux value}

\item{lux_2}{second measured lux value}

\item{n}{number of individuals in sample}

\item{population_df}{a tibble representing a virtual experiment}
}
\value{
a list of two sets of measured melatonin values
}
\description{
These samples can either be from different individuals (for a between-type
study) or from the same individuals measured twice (for a within-type study).
}
