% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{generate_two_samples_at_one_lux}
\alias{generate_two_samples_at_one_lux}
\title{Create two samples of melatonin suppression observations from an intervention type experiment}
\usage{
generate_two_samples_at_one_lux(is_between, lux, n, population_df)
}
\arguments{
\item{is_between}{a Boolean indicating whether experiment is within or between type}

\item{lux}{the lux value at which to conduct the comparison}

\item{n}{number of individuals in sample}

\item{population_df}{a tibble representing a virtual experiment}
}
\value{
a list of two sets of measured melatonin values
}
\description{
Create two samples of melatonin suppression observations from an intervention type experiment
}
