% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logistic_2}
\alias{logistic_2}
\title{Two-parameter dose-response logistic curve}
\usage{
logistic_2(log_lux, p1, p2)
}
\arguments{
\item{log_lux}{a log lux value}

\item{p1}{the log of the ed50 lux}

\item{p2}{a positive shape parameter}
}
\value{

}
\description{
Returns a melatonin suppression response for a given level of log_lux
}
\examples{
# below returns 0.5 since log_lux=log_ed50
logistic_2(10, 10, 5)
}
