% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose-response.R
\name{logistic_noise}
\alias{logistic_noise}
\title{Generates noisy dose response data at a range of (log-)luxes}
\usage{
logistic_noise(
  sigma,
  p1,
  p2,
  log_lux = log10(c(10, 30, 50, 100, 200, 400, 2000))
)
}
\arguments{
\item{sigma}{a noise level (represents additive Gaussian noise on logit scale)}

\item{p1}{the log of the ed50 lux}

\item{p2}{a positive shape parameter}

\item{log_lux}{a vector of log lux values which defaults to log10(c(10, 30, 50, 100, 200, 400, 2000))}
}
\value{
a tibble of paired lux and melatonin suppression values
}
\description{
Adds noise on the logit scale, resulting in more realistic (noisy) data for
a given patient.
}
\examples{
logistic_noise(1, 1.3, 4.5)
}
\seealso{
\link{noise_logit}
}
