% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{noise_logit}
\alias{noise_logit}
\title{Adds noise to a data point on the logit scale}
\usage{
noise_logit(y, sigma)
}
\arguments{
\item{y}{A melatonin suppression value between 0 and 1}

\item{sigma}{Scale of Gaussian noise}
}
\value{
A noisy version of y bound between 0 and 1
}
\description{
Adds Gaussian noise to a point on the logit scale:
\deqn{logit(y_noise) = logit(y) + e}
where:
\deqn{e \sim normal(0, sigma)}
}
\details{
Then we convert back to the normal scale via a logistic transform.
}
