% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p1_p2_regression_draws}
\alias{p1_p2_regression_draws}
\title{Posterior parameter draws for regression model of (log-p2) on p1.}
\format{A list with 4000 parameter draws for each of:
\describe{
\item{alpha}{intercept parameter}
\item{beta}{slope parameter}
\item{sigma0}{noise intercept parameter}
\item{sigma1}{noise slope parameter}
...
}}
\usage{
p1_p2_regression_draws
}
\description{
A Stan model was used to fit a model to the Phillips et al. (2017) point
estimates. The model was of the form:
\deqn{log p2 \sim normal(alpha + beta p1, sigma0 + sigma1 * p1)}
}
\references{
Phillips et al., (2017). High sensitivity and interindividual variability
in the response of the human circadian system to evening light. PNAS.
}
\keyword{datasets}
