% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_doseresponse}
\alias{plot_doseresponse}
\title{Plots simulated data and dose-response curves for a population}
\usage{
plot_doseresponse(experiment_df, plot_curves = TRUE)
}
\arguments{
\item{experiment_df}{a simulated population}

\item{plot_curves}{a Boolean indicating whether to plot the true dose-response curves}
}
\value{
a ggplot object
}
\description{
Plots simulated data and dose-response curves for a population
}
\examples{
# generate dose-response data for n=200 individuals measured at four lux levels
experimental_data <- virtual_experiment(n=200, lux=c(1, 10, 100, 1000))

# plot data
plot_doseresponse(experimental_data)
}
