% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose-response.R
\name{sample_p1_p2}
\alias{sample_p1_p2}
\title{Samples values of pairs of logistic_2 dose-response parameters using models fit to Phillips et al. (2017) estimates}
\usage{
sample_p1_p2(n, alpha, beta, sigma0, sigma1, cdf_inv)
}
\arguments{
\item{n}{the number of individual dose-response curves to generate}

\item{alpha}{posteriors draws for the intercept in the regression of log_p2 on p1}

\item{beta}{posterior draws for slope parameter in the regression of log_p2 on p1}

\item{sigma0}{posterior draws for the constant noise term in the regression of log_p2 on p1}

\item{sigma1}{posterior draws for the heteroscedastic noise term in the regression of log_p2 on p1}

\item{cdf_inv}{inverse cumulative density function for distribution of p1 observed in Phillips et al., (2017)}
}
\value{
a tibble of p1 and p2 values for each individual dose-response curve
}
\description{
Creates a set of dose-response curves -- one per individual -- supposed to mimic participants in an experiment.
This is based on Bayesian models fit to parameter estimates from Phillips et al., (2017).
}
