% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose-response.R
\name{sample_sigma}
\alias{sample_sigma}
\title{Samples a value of sigma for the logit noise process}
\usage{
sample_sigma()
}
\value{
a positive value
}
\description{
The sampled value is based on fitting a gamma distribution to
estimates of the noise values using RMSEs from Phillips et al. (2017).
The gamma distribution was fit using Stan, so estimates here incorporate
posterior uncertainty in the parameter values.
}
