% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sigma_fit_draws}
\alias{sigma_fit_draws}
\title{Posterior parameter draws for model of sigma}
\format{A list with 4000 parameter draws for each of:
\describe{
\item{a}{gamma parameter 1 in Stan definition of distribution}
\item{b}{gamma parameter 2 in Stan definition of distribution}
}}
\usage{
sigma_fit_draws
}
\description{
The model (a gamma distribution) was fit in Stan.
}
\keyword{datasets}
