% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{t_test}
\alias{t_test}
\title{Performs t test between two sets of values}
\usage{
t_test(is_between, vals)
}
\arguments{
\item{is_between}{a Boolean indicating whether experiment is within or between type}

\item{vals}{a named list of 'vals_1' and 'vals_2', each of the same length}
}
\value{
a list with class "htest" resultant from performing a t.test
}
\description{
Ensures that tests are conducted assuming paired values if required
}
