% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose-response.R
\name{treated_p1}
\alias{treated_p1}
\title{Converts p1 assuming treatment that changes ed50 by a multiplier}
\usage{
treated_p1(multiplier, old_p1)
}
\arguments{
\item{multiplier}{indicates new ed50 = old ed50 * multiplier}

\item{old_p1}{p1 value before treatment}
}
\value{
a new p1 value
}
\description{
Converts p1 assuming treatment that changes ed50 by a multiplier
}
