% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose-response.R
\name{valid_individual}
\alias{valid_individual}
\title{Generates an individual with reasonable ed25 and ed75 values.}
\usage{
valid_individual(
  thresh_25,
  thresh_75,
  eds_25,
  eds_75,
  p1_distribution_parameters,
  p2_distribution_parameters
)
}
\arguments{
\item{thresh_25}{lower bound on simulated ed25 vs observed ed25 (bound calculated as thresh_25 * observed)}

\item{thresh_75}{upper bound on simulated ed75 vs observed ed75 (bound calculated as thresh_75 * observed)}

\item{eds_25}{vector of ed25s from estimates from Phillips et al. (2017)}

\item{eds_75}{vector of ed75s from estimates from Phillips et al. (2017)}

\item{p1_distribution_parameters}{a named list of items used to characterise the p1 distribution:
cdf_inv_full, normal_sigma, weight_full. Here cdf_inv is the inverse-cdf
characterising the distribution of p1 distribution from Phillips et al. (2017); normal_sigma is the standard
deviation of the normal distribution used as the narrow mixture component; weight_full is the weight given to the
p1_distribution derived from Phillips et al. (2017).}

\item{p2_distribution_parameters}{a named list of items used to characterise the log_p2|p1 distribution. It comprises:
"alpha" posteriors draws for the intercept in the regression of log_p2 on p1; "beta" posterior draws for slope
parameter in the regression of log_p2 on p1; "sigma0" posterior draws for the constant noise term in the regression of log_p2 on p1;
"sigma1" posterior draws for the heteroscedastic noise term in the regression of log_p2 on p1;
"weight" a number (0<=weight<=1) used to control the variance reduction in individual heterogeneity.}
}
\value{
a tibble containing dose-response parameters for an individual
}
\description{
Generates an individual with reasonable ed25 and ed75 values.
}
