% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose-response.R
\name{virtual_population}
\alias{virtual_population}
\title{Generates a population of individual dose-response curves}
\usage{
virtual_population(n, thresh_25, thresh_75, weight_p1, weight_p2)
}
\arguments{
\item{n}{number of individual dose-response curves to generate}

\item{thresh_25}{lower bound on simulated ed25 vs observed ed25 (bound calculated as thresh_25 * observed)}

\item{thresh_75}{upper bound on simulated ed75 vs observed ed75 (bound calculated as thresh_75 * observed)}

\item{weight_p1}{the weight (0<=weight<=1) of the purely empirical distribution of p1 derived from Phillips et al. (2017)}

\item{weight_p2}{the weight (0<=weight<=1) specifying the factor to suppress variation in p2_log|p1}
}
\value{
a tibble with n rows with pairs of logistic_2 dose-response parameters
}
\description{
Generates a population of individual dose-response curves
}
