% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose-response.R
\name{virtual_population_generator}
\alias{virtual_population_generator}
\title{Generates a population of individual dose-response curves}
\usage{
virtual_population_generator(n, thresh_25, thresh_75)
}
\arguments{
\item{n}{number of individual dose-response curves to generate}

\item{thresh_25}{lower bound on simulated ed25 vs observed ed25 (bound calculated as thresh_25 * observed)}

\item{thresh_75}{upper bound on simulated ed75 vs observed ed75 (bound calculated as thresh_75 * observed)}
}
\value{
a tibble with n rows with pairs of logistic_2 dose-response parameters
}
\description{
Generates a population of individual dose-response curves
}
