% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose-response.R
\name{valid_individual}
\alias{valid_individual}
\title{Generates an individual with reasonable ed25 and ed75 values.}
\usage{
valid_individual(
  thresh_25,
  thresh_75,
  eds_25,
  eds_75,
  alpha,
  beta,
  sigma0,
  sigma1,
  cdf_inv
)
}
\arguments{
\item{thresh_25}{lower bound on simulated ed25 vs observed ed25 (bound calculated as thresh_25 * observed)}

\item{thresh_75}{upper bound on simulated ed75 vs observed ed75 (bound calculated as thresh_75 * observed)}

\item{alpha}{posteriors draws for the intercept in the regression of log_p2 on p1}

\item{beta}{posterior draws for slope parameter in the regression of log_p2 on p1}

\item{sigma0}{posterior draws for the constant noise term in the regression of log_p2 on p1}

\item{sigma1}{posterior draws for the heteroscedastic noise term in the regression of log_p2 on p1}

\item{cdf_inv}{inverse cumulative density function for distribution of p1 observed in Phillips et al., (2017)}
}
\value{

}
\description{
Generates an individual with reasonable ed25 and ed75 values.
}
