% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose-response.R
\name{virtual_experiment}
\alias{virtual_experiment}
\title{Generates data from a virtual experiment measuring individual dose-response curves}
\usage{
virtual_experiment(
  n,
  lux = c(10, 30, 50, 100, 200, 400, 2000),
  thresh_25 = 0.5,
  thresh_75 = 1.5
)
}
\arguments{
\item{n}{the number of individual dose-response curves to generate}

\item{lux}{a vector of lux values which defaults to c(10, 30, 50, 100, 200, 400, 2000)}

\item{thresh_25}{lower bound on simulated ed25 vs observed ed25 (bound calculated as thresh_25 * observed) which defaults to 0.5}

\item{thresh_75}{upper bound on simulated ed75 vs observed ed75 (bound calculated as thresh_75 * observed) which defaults to 1.5}
}
\value{
a tibble containing 'measured' dose-response melatonin curves at each lux value for each individual
}
\description{
The model used to generate these experiments comprises two elements: a model representing
the underlying dose-response curves (which is based on a two parameter logistic); and a
model of typical experimental error in these measurements. The model was fit using estimates
presented in Phillips et al., (2017).
}
